+++ python/GafferSceneTest/__init__.py
@@ -180,6 +180,7 @@ from .RenderPassWedgeTest import RenderPassWedgeTest
 from .RenderAdaptorTest import RenderAdaptorTest
 from .MeshTessellateTest import MeshTessellateTest
 from .RenderPassTypeAdaptorTest import RenderPassTypeAdaptorTest
+from .RenderPassShaderTest import RenderPassShaderTest
 
 from .IECoreScenePreviewTest import *
 from .IECoreGLPreviewTest import *